Resolved conflict in App.jsx
