+++ src/components/MultiStepForm.jsx
@@ -1,4 +1,3 @@
-/* eslint-disable react/prop-types */
 /**
  * This component represents a multi-step form that collects user information such as favorite music and favorite animal */
